Create Example .exe
This is an example run of the program. When the user hits 'run', they are given a team of creatures, and an enemy with a team of creatures. The creatures automatically attack, swap out for another team member, or rest depending on their health and the health of the rest of the team. The game is over when all members of a team reach 0 health.